Refactor document flow to separate layers from each other
