+++ src/main/kotlin/no/elhub/auth/data/persist/tables/AuthorizationScopeTable.kt
@@ -1,4 +1,4 @@
-package no.elhub.auth.data.exposed.tables
+package no.elhub.auth.data.persist.tables
 
 import no.elhub.auth.domain.shared.AuthorizationResourceType
 import no.elhub.auth.domain.shared.PermissionType
