+++ src/main/kotlin/no/elhub/auth/data/persist/tables/PGEnum.kt
@@ -1,4 +1,4 @@
-package no.elhub.auth.data.exposed.tables
+package no.elhub.auth.data.persist.tables
 
 import org.postgresql.util.PGobject
