+++ src/main/kotlin/no/elhub/auth/domain/document/DocumentRepository.kt
@@ -1,0 +1,8 @@
+package no.elhub.auth.domain.document
+
+import java.util.UUID
+
+interface DocumentRepository {
+    fun insertDocument(doc: AuthorizationDocument)
+    fun getDocumentFile(id: UUID): ByteArray?
+}
