+++ src/main/kotlin/no/elhub/auth/domain/shared/AuthorizationScope.kt
@@ -2,8 +2,8 @@ package no.elhub.auth.domain.shared
 
 import kotlinx.datetime.Instant
 import kotlinx.serialization.Serializable
-import no.elhub.auth.data.exposed.tables.AuthorizationGrantTable
-import no.elhub.auth.data.exposed.tables.AuthorizationScopeTable
+import no.elhub.auth.data.persist.tables.AuthorizationGrantTable
+import no.elhub.auth.data.persist.tables.AuthorizationScopeTable
 import org.jetbrains.exposed.sql.ReferenceOption
 import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.javatime.timestamp
@@ -31,7 +31,6 @@ enum class PermissionType {
     ReadAccess
 }
 
-
 object AuthorizationGrantScopeTable : Table("auth.authorization_grant_scope") {
     val authorizationGrantId = uuid("authorization_grant_id")
         .references(AuthorizationGrantTable.id, onDelete = ReferenceOption.CASCADE)
@@ -40,5 +39,3 @@ object AuthorizationGrantScopeTable : Table("auth.authorization_grant_scope") {
     val createdAt = timestamp("created_at").clientDefault { java.time.Instant.now() }
     override val primaryKey = PrimaryKey(authorizationGrantId, authorizationScopeId)
 }
-
-
