+++ src/main/kotlin/no/elhub/auth/presentation/AuthorizationGrantRoute.kt
@@ -7,15 +7,15 @@ import io.ktor.server.routing.Route
 import io.ktor.server.routing.get
 import io.ktor.server.routing.route
 import io.ktor.server.util.url
+import no.elhub.auth.domain.grant.AuthorizationGrantHandler
+import no.elhub.auth.domain.grant.AuthorizationGrantProblem
 import no.elhub.auth.presentation.config.ID
 import no.elhub.auth.presentation.model.errors.ApiError
 import no.elhub.auth.presentation.model.errors.ApiErrorJson
-import no.elhub.auth.presentation.model.toGetAuthorizationGrantScopeResponse
-import java.util.UUID
-import no.elhub.auth.domain.grant.AuthorizationGrantHandler
-import no.elhub.auth.domain.grant.AuthorizationGrantProblem
 import no.elhub.auth.presentation.model.toGetAuthorizationGrantResponse
+import no.elhub.auth.presentation.model.toGetAuthorizationGrantScopeResponse
 import no.elhub.auth.presentation.model.toGetAuthorizationGrantsResponse
+import java.util.UUID
 
 fun Route.grants(grantHandler: AuthorizationGrantHandler) {
     route("") {
