+++ src/main/kotlin/no/elhub/auth/presentation/AuthorizationRequestRoute.kt
@@ -10,17 +10,17 @@ import io.ktor.server.routing.Route
 import io.ktor.server.routing.get
 import io.ktor.server.routing.post
 import io.ktor.server.util.url
-import no.elhub.auth.presentation.config.ID
-import no.elhub.auth.presentation.model.errors.ApiError
-import no.elhub.auth.presentation.model.errors.ApiErrorJson
-import java.util.UUID
 import no.elhub.auth.domain.request.AuthorizationRequestHandler
 import no.elhub.auth.domain.request.AuthorizationRequestProblemById
 import no.elhub.auth.domain.request.AuthorizationRequestProblemCreate
 import no.elhub.auth.domain.request.AuthorizationRequestProblemList
+import no.elhub.auth.presentation.config.ID
 import no.elhub.auth.presentation.model.PostAuthorizationRequestPayload
+import no.elhub.auth.presentation.model.errors.ApiError
+import no.elhub.auth.presentation.model.errors.ApiErrorJson
 import no.elhub.auth.presentation.model.toGetAuthorizationRequestResponse
 import no.elhub.auth.presentation.model.toGetAuthorizationRequestsResponse
+import java.util.UUID
 
 fun Route.requests(requestHandler: AuthorizationRequestHandler) {
     get {
