+++ src/main/kotlin/no/elhub/auth/presentation/Validators.kt
@@ -2,9 +2,9 @@ package no.elhub.auth.presentation
 
 import arrow.core.Either
 import arrow.core.raise.either
-import no.elhub.auth.presentation.model.errors.ApiError
-import no.elhub.auth.presentation.model.PostAuthorizationRequestPayload
 import no.elhub.auth.domain.request.AuthorizationRequest
+import no.elhub.auth.presentation.model.PostAuthorizationRequestPayload
+import no.elhub.auth.presentation.model.errors.ApiError
 import java.util.UUID
 
 /**
