+++ src/test/kotlin/no/elhub/auth/data/persist/ExposedDocumentRepository.kt
@@ -1,22 +1,22 @@
-package no.elhub.auth.data.exposed
+package no.elhub.auth.data.persist
 
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldNotBe
-import java.time.LocalDateTime
-import java.util.*
-import no.elhub.auth.data.exposed.repositories.AuthorizationDocumentRepository
-import no.elhub.auth.data.exposed.tables.AuthorizationDocumentScopeTable
-import no.elhub.auth.data.exposed.tables.AuthorizationScopeTable
+import no.elhub.auth.data.persist.repositories.ExposedDocumentRepository
+import no.elhub.auth.data.persist.tables.AuthorizationDocumentScopeTable
+import no.elhub.auth.data.persist.tables.AuthorizationScopeTable
 import no.elhub.auth.domain.document.AuthorizationDocument
 import no.elhub.auth.extensions.PostgresTestContainerExtension
 import org.jetbrains.exposed.sql.Database
 import org.jetbrains.exposed.sql.selectAll
 import org.jetbrains.exposed.sql.transactions.transaction
+import java.time.LocalDateTime
+import java.util.*
 
-class AuthorizationDocumentRepositoryTest :
+class ExposedDocumentRepository :
     FunSpec({
         extensions(PostgresTestContainerExtension)
-        val repository = AuthorizationDocumentRepository
+        val repository = ExposedDocumentRepository()
 
         beforeSpec {
             Database.connect(
@@ -44,10 +44,10 @@ class AuthorizationDocumentRepositoryTest :
                     )
 
                 // When
-                AuthorizationDocumentRepository.insertDocument(document)
+                repository.insertDocument(document)
 
                 // Then
-                val documentExists = AuthorizationDocumentRepository.getDocumentFile(document.id)
+                val documentExists = repository.getDocumentFile(document.id)
                 documentExists shouldNotBe null
 
                 val authorizationDocumentScopeRow =
